`create-if` may fail for self-assembling zones (#3222)
Part of the work to set MTUs involved setting the IP interface MTU for
control vnics inside the zone they're used in. The problem is
`create-if` will fail if the vnic being used inside a zone already has
an interface in that zone. This occurs when the zone is self-assembling,
meaning some software inside the zone is responsible for bringing up a
listen address for its service(s). What would happen is sled-agent's
`create-if` would race with something inside the zone running
`create-addr`: these zones would eventually come up after racing for a
long time but were delayed.

Fix this by detecting if a zone is self-assembling, and eating
`create-if` error if it is. `create-if` still needs to be run so that
the `set-ifprop` commands also run as part of setting MTU don't fail due
to a missing interface.