Adjust chrony configuration based on experience on rack2
Fixes #3137

If RSS is started before external connectivity is established,
it is possible for non-boundary NTP servers to get stuck in
1987 and they will take a long time to catch up from there
as they cannot set the clock slew beyond 32,500 ppm.

I'm also enabling additional logging and log rotation here
so that we can get better visibility into this in the future
and I am configuring the leap second clock slew parameters.