sled agent: don't publish the Destroyed instance state
Nexus reserves the "Creating" and "Destroyed" instance states to refer to
instances that it (Nexus) is in the process of creating and that it has deleted.
Propolis uses these states to refer to VMMs that it has created and destroyed.
Sled agent is responsible for correcting this impedance mismatch by turning
Propolis's "per-VMM" states into appropriate Nexus instance states.

Fix two places where sled agent was erroneously publishing the Destroyed state.
Adjust some types to make it harder to make this mistake in the future.