[sled-agent] Encrypt U.2 zpool top-level datasets (#3007)
Encryption of U.2 devices relies on key generated by the `KeyManager`
introduced in https://github.com/oxidecomputer/omicron/pull/2990. The
`KeyManager` is parameterized by a
`SecretRetriever`, which provides input key material (IKM) that the
`KeyManager` can use to derive keys. In this commit we introduce a
`LocalSecretRetriever` which provides secrets local to a sled. Right
now the secret is hardcoded, but we intend to either store random
secrets on the M.2s or retrieve them from VPD data in the short
term. Longer term, we will replace the `LocalSecretRetriever` with a
`TrustQuorumSecretRetriever` which will provide a rack secret recomputed
from trust quorum key shares.

The `KeyManger` has been  extended to run in an tokio task and process
requests for from different "requesters". Each requester only has the
ability to request keys that it requires to do its job. Currently there
is a single requester, `StorageKeyRequester`, that is owned by the
`StorageManager` which loans it out to `sled_hardware::Disk` to request
keys on demand.

`sled_hardware::Disk` requests keys for both creation of encrypted
datasets and mounting of existing encrypted datasets. In order to
actually allow ZFS to use the generated keys, it creates keyfiles inside
a RAMDisk. Attempts are made to limit the lifetime of these files as
well as accidental leakage of secrets.

Changes have also been made to `illumos_utils::Zfs` to support
encryption. To support future secret reconfiguration with trust quorum
we store an additional `epoch` property on each encrypted dataset, which
allows the `KeyManager` to retrieve the right vesion of the secret from
a `SecretRetriever`.