[trust-quorum] Start implementing version 0 (#3167)
This is the start of a stripped down version of the trust quorum
bootstore for use in our MVP racks. We define "schemes", such that each
version of trust quorum can get stronger security wise over time, and we
can upgrade into new schemes. This change starts to implement `bootstore
scheme version 0`, where all important parameters exist in a self
describing zero-sized struct, `V0Scheme`, that can be debug printed for
informational purposes, such as logging in sled-agent startup.

This commit implements the first part of the V0 scheme, which is the
necessary datastructures around trust quorum. As these structures are
all security sensitive, none of them are going to be leaving the
bootstore logic itself. The sole exception to this is that the rack
secret will be handed off to the `KeyManager` in sled-agent in order to
derive disk encryption keys. This will come in a follow up commit.

There are two primary datastructures introduced here:
* `RackSecret` - A thin wrapper around shamir secret sharing using
Curve25519
* `SharePkg` - A container of information distributed to each of the
initial sleds that are part of the rack during rack initialization

`RackSecret` itself will never be stored on disk, and `SharePkg` and its
analog `LearnedSharePkg` for sleds that were not part of the original
rack membership, will be stored on the M.2s on each sled in a rack. We
plan to extract the `Ledger` from sled-agent into `omicron-common` and
use that to synchronize writing across both M.2s. We include a
`SharePkgHeader` that can be used to distinguish what type of share
header is stored on the M.2s so that we can properly deserialize them.
We expect this structure to be remain backwards compatible indefinitely.

For each sled that is part of the initial group, we include a set of
encrypted shares in its `SharePkg`. This is to allow other sleds to get
added to the rack and get their own key-shares over time. These shares
are encrypted so that an attacker stealing fewer than threshold sleds
(or their M.2s) cannot recover the rack secret. When a new sled is added
it will ask for a share from one of the existing sleds which will return
it a share and mark it as used on its M.2s. In order to return a share,
an existing sled will have to retrieve enough shares to reconstruct the
rack secret and decrypt the shares it has in its `SharePkg`. If it does
not have any shares remaining, the asking sled will have to go and ask
another sled. This is a best-effort mechanism with known security and
reliability issues. However, it allows us to ship a version of trust
quorum that will be able to survive sled additions long enough for us to
implement trust quorum V1.

In scheme version 0, shares and packages will be distributed between
sleds over unencrypted tcp links rather than sprockets sessions. We also
will distribute them to anyone on the bootstrap network who asks. This
is a strategy designed to allow our primary goal of preventing offline
data recover from stolen U.2 devices at MVP. It aligns well with our
current control-plane limitations around unencrypted transport that we
plan to fix over time.

There will be a few follow up commits in order to complete scheme v0 and
integrate it into our system:

1. Augment the bootstore with the V0 protocol over TCP channels that
allows `RackSecret` reconstruction and `SharePkg` distribution
2. Integrate this protocol into the bootstrap agent with calls into the
bootstore and remove the ad-hoc code that currently exists.
3. Add a `StorageKeyRequester` for the the v0 trust quorum so that the
`KeyManager` can generate disk encryption keys using the `RackSecret` as
input key material (IKM)

As part of this commit, almost all the previous bootstore code was
deleted. It is obsolete, as we are planning to use a quorum based,
rather than 2 phase commit based strategy for reconfiguration. We also
plan to separate the persistence layer from the stored data and not use
sqlite transactions as the implementation of invariants for our
protocol. This should make testing easier, and allow us a choice of
using async or sync connections to sqlite, or even abandoning sqlite in
favor of flat files or another persistence strategy. The new protocol
still needs to be written as an update to RFD 355.

Fixes #1513 and #3158