instance create saga: create new DAG for each test iteration
Modify the instance create saga's unwind tests so that they create a new saga
DAG for every iteration of the test (i.e., for each distinct node in the DAG
into which a failure can be injected). This ensures that every test iteration
uses a DAG with a unique instance ID, which is needed to ensure that each
iteration can actually reach the node into which it intends to inject an error.

Add a step to the end of the affected tests that verifies that a previously-
created-but-unused DAG can still run to completion after all the error injection
tests have been run.