sled agent: improve instance state update retry loop (#3256)
Refine the instance state update retry loop as follows:

- Check return codes from Nexus much more carefully. In particular,
treat client errors as permanent errors that should stop the update
loop. This avoids looping forever if an instance is deleted (in Nexus)
when its Propolis is stopped but not yet fully torn down (see #3230 for
a description of how this occurs).
- Log information about the updates that are being pushed. Also be sure
to log when an update attempt fails permanently.
- Switch the retry policy from "local" (meant for operations that don't
leave the sled) to the much less aggressive "internal service." The
'caller is invoking other internal services' semantics are especially
appropriate here because post-migration state updates can cause Nexus to
do a lot of work that contacts other rack services.
- Don't bail out of the instance state monitor task merely because it
failed to push an update to Nexus, since this will cause future updates
from the task's Propolis to be ignored.

Tested: cargo test; ran an Omicron dev cluster and created/destroyed a
few instances; reproduced #3230 and #3231 and observed that the
notification worker stops retrying if an intervening instance
destruction causes a notification to return 404 Not Found.

Fixes #3230. Fixes #3231.