[wicketd] Track trampoline images by hash instead of ID
If the user uploads a TUF repository with a different trampoline image
but the same metadata (version, in particular), wicketd would _not_
treat that as a new trampoline image when checking whether it needed to
be sent to MGS upon starting an update.

We now do this comparison by hash of the image instead of by metadata,
ensuring a new image will always be sent to MGS even if the metadata is
unchanged.

Fixes https://github.com/oxidecomputer/omicron/issues/3282.