Use native helios-build image name tokens
Rather than retrieving the package repository ourselves
and parsing the OS hash out of it, we can make use of
the new token feature in helios-build and have it populate
the OS hash during build. This also closes the small window
that could result in the image name having an older hash
than the image.