instance create saga: create new DAG for each test iteration (#3309)
Modify the instance create saga's unwind tests so that they use a distinct
instance ID for every iteration of the test (i.e., for each distinct node
into the DAG into which a failure can be injected). Reusing the same
instance ID causes the create saga to fail repeatedly at the "create
instance record" step instead of at the later steps the tests intend.

To help check this, add a test function that executes a saga and returns
the raw Steno saga result, then use that to check that the sagas failed at
the expected nodes.