Use `zone_enter` instead of `zlogin` to avoid shell expansion. (#3316)
- Fixes #3216
- For running zones, use the `Command::pre_exec` hook to call
`zone_enter()` between `fork()` and `exec()`. This lets us run the
command directly in the zone, without using `zlogin`. The latter runs
all its arguments through a shell in the form of `sh -c`, which means
they're subject to expansion. Using `zone_enter()` instead avoids any
expansion.
- Add `Zones::id()` method, which mostly exists for testing. `mockall`
requires us to specify the return value for the mock methods, but the
`Zone` type we're using has private fields, so it can't be easily faked.