Fix SAML login redirect (#3318)
When you get logged out in the console or get linked to a page while not
logged in (like `/device/verify` for authenticating in the CLI), we want
to send you somewhere you can log in and, after you succeed, we want to
send you back where you were trying to go. In order to do that, we need
to keep track of the original target page and hold onto it through the
login process.

With local login, this is easy: in the case of a server-side 401, we
take the current path, and redirect to
`/login/{silo}/local?redirect_uri={path}`, which loads a console page
that handles username/password login through a background API call. When
the login POST comes back successful, we pull the target path from the
`state` query param and nav to it client-side ([console
source](https://github.com/oxidecomputer/console/blob/d52e87d69877e9dd63254dfa44ef8c284a2d3518/app/pages/LoginPage.tsx#L29-L36)).
(We need to make sure you can't link someone to
`oxide-rack.com/login/{silo}/local?redirect_uri=https://evil.com/virus`
— see to-do list below.) A client-side 401 is almost the same, except
there's one more step at the beginning where we nav to
`/login?redirect_uri={current_path}`, and then that handler sends you to
local or SAML login as appropriate, making sure to pass along the
`state` param (very important!).

With SAML login it's slightly more complicated because we need to send
the target URL over to the IdP using SAML's `RelayState` concept and
they will send it back with the login POST. Most of that is already in
place, so the main thing this change does is pull the target URL from
the `state` query param server side instead of the `referer` header and
stick it in the `RelayState` we get back from the IdP on login.

### What's in here

- [x] Move existing redirect to IdP logic from
`/login/{silo}/saml/{provider}` to
`/login/{silo}/saml/{provider}/redirect`,
`/login/{silo}/saml/{provider}` now serves console landing page
- [x] Get post-login redirect URL from `state` query param instead of
`referer` header
- [x] Rename `state` param to `redirect_uri` and use that in both the
query params and the RelayState object
- [x] Validate that `redirect_uri` is a relative path to prevent someone
being given a login link with an arbitrary URL
45e2fbe93bef09a1c04ecc8ccee79b221ef3d5ed
- [x] Better tests, probably
- ~~Change state param from path to full URL per this comment~~ Relative
URIs are valid in a 302 response Location header ([RFC
7231](https://www.rfc-editor.org/rfc/rfc7231#section-7.1.2)). I wrote
that comment because I thought the IdP would be returning a redirect to
Nexus, but it seems they post to Nexus from the browser, and browsers
resolve a relative Location against the request URI

https://github.com/oxidecomputer/omicron/blob/8a1fbdfb8244d232274ee03d82a47b3186b89cf4/nexus/src/external_api/console_api.rs#L593-L599

### Why not the `referer` header

`referer` certainly does sound like the natural choice, but browser
behavior around it is inconsistent, especially in the case of multiple
redirects. When you run into a 401 client-side (like if your session
expired), the console uses `window.location.assign` to send the user to
`/login` on Nexus, which will return a redirect to either
`/login/{silo}/local` or `/login/{silo}/saml/{idp}` depending on the
silo.


https://github.com/oxidecomputer/omicron/blob/8a1fbdfb8244d232274ee03d82a47b3186b89cf4/nexus/src/external_api/console_api.rs#L559-L577

In theory the original `referer` should be passed through those
redirects and accessible on the final page, but I don't see it when I
test it manually. I might do more testing, but overall I don't trust it.
It's browser-specific, who knows what could happen. For now I think it's
more reliable to use a query param we fully control.