Support for nexus sub tags (#3235)
Using the format `{tag}/{subtag}` we can nest operations. This solves
the issue of tags that contain so many operations that it gets
[difficult to
navigate](https://github.com/oxidecomputer/docs/pull/96#issuecomment-1563226809).

I've updated the URLs, these don't yet work but I'll add an issue on the
docs site to get these matched with the tags.
The docs site will also need updating to render the nested operations,
but I'll address that after this is merged.

I'd initially went with `networking/{subtag}` — but figured
`vpcs/{subtag}` was closer to the existing API.

This will break the traits: `ClientSystemExt` and `ClientVpcsExt`. I've
updated usage of those in this repo and done a check across the org but
there's a good chance I might have missed something.

I've tried to work on the descriptions, will need someone to vet those
to make sure they are accurate.