Generate dendrite/maghemite PRs (#3332)
Implements PR generation for #3259 .

There is a GitHub action workflow per dependency for performing updates,
but both of them perform the same steps in this case. Specifically they
will:

1. Checkout the defined integration branch for the dependency (or create
one if it does not exist)
2. Merge main in to the integration branch 
3. Run the appropriate `update_*.sh` script
4. Commit and push those changes to the integration branch
5. Generate a new PR if there is not yet an open PR from the integration
branch to main. If there is an open PR already, then it will update the
PR instead.

PRs will have title and body that look like:
```
Title: Update dendrite to 762bcd6
Body:
Updated dendrite API version to 762bcd6
```

Currently the version is extracted from `tools/*_openapi_version`.

The actions are set up to run in response to pushes to the main branches
of the dendrite and maghemite repositories. There is a one hour delay
(during which new commits will auto-cancel previously scheduled actions)
set to allow for the respective commits to finished their builds. In a
future state this could be improved to instead listen for a build
completion event to trigger the update action.