Only update relevant ports when updating firewall rules on sled. (#3352)
Firewall rules are always pushed per-VPC but we were previously attempting to apply said rules to every OPTE port on a sled. This would cause any ports not tied to the specific VPC to have their firewall rules cleared out. Filtering to just the relevant ports before attempting to apply the rules addresses this.

Fixes #3336