[wicket] Allow CLI-style RSS configuration (#3326)
This is the first of 4 (ish?) PRs building toward configuring and
running RSS via wicket. It adds a handful of endpoints to wicketd for
various RSS configuration settings, and a preliminary CLI interface with
which a user can supply those settings.

This isn't 100% correct or complete and will be tweaked in some of the
followup PRs, but the majority of it is close enough to right that I
think it's ready for review on its own.