Clarify the CTE around region allocation (#3374)
Some minor docs improvements, as I dug through this codebase recently:

- I wanted to clarify, "Does this CTE only provision Crucible datasets
to U.2s, and not M.2s?". The answer to this question is: "yes, but
implicitly decided when we provision the datasets themselves". Added a
comment to clarify this during region selection.
- Additionally, when referencing a virtual table from an earlier `WITH
candidate_zpools AS ...` arm of the CTE, make that reference explicit.
This is a no-op change, but it clarifies the dependencies within the
CTE.