WIP: Introduce bootstore FSMs for trust quorum V0
This commit implements and tests the bootstore protocol for trust quroum
version 0 (aka low-rent-trust-quorum). Rack initialization and sled
unlock is written and tested.

The code is WIP because there are still some TODOs related to
functionality and testing around adding new sleds to the cluster. This
commit is also enormous, and I'd like to add documentation for the code
structure and test strategy prior to marking it ready for review. That
being said, the code is close enough to start reviewing and it follows
the description in the overview meet video.

Note, that importantly, and very deliberately, this code performs no IO!
There will be a thin async layer on top for managing tcp connections and
proxying messages. Additionally, there will be a small persistence layer
for when the `Output` indicates persistance is needed. It's anticipated
that we will extract and use the `Ledger` currently in sled-agent for
this as it already writes to both M.2s. The ledger may need some small
mods for this to work.