[wicketd] Add periodic check for sled agent bootstrap addrs (#3327)
RSS wants to know the bootstrap addresses of the sleds to include; this
PR teaches wicketd how to find them. It also updates the wicket TOML
config introduced in #3326 to include the sled's bootstrap address (if
known) in a comment in the sled list; I'm unsure if this is something
we'll want to keep once customers are running RSS, but it seems handy
for debugging for now.

Builds on #3326.