Change test model to operate in lockstep with fsms
We no longer try to aggregate expected output from
the model by updating it via `Action`s. Instead
we call it 1:1 with each calback used on the FSMs
so we can just compare outputs directly.

Also, found and fixed a bunch of bugs and added
new functionality.

Proptests ran in a loop for 10 min with no failures.