[wicket/wicketd] Run RSS from Wicket (#3387)
The bulk of code in this PR is the new `rack_setup` wicket pane, which
includes a bunch of popup-related things (which are all fairly verbose
for now; I followed what the update pane does structurally, and am very
thankful it exists!). On a fresh (uninitialized and unconfigured)
system, the pane should look roughly like this, although hopefully not
with `UNKNOWN` bootstrap addresses (those are because these screenshots
are from a simulated dev environment that doesn't have a bootstrap
network):


![pr0](https://github.com/oxidecomputer/omicron/assets/1435635/6ce9d4b0-4290-48d8-8794-63552902e184)

If we attempt to start an update, we'll get a confirm popup; pressing
"Yes" will fail with the error from wicketd (in this case, "you need to
configure things!"):


![pr1](https://github.com/oxidecomputer/omicron/assets/1435635/46152156-3313-4eec-b0d8-221e43f4cf81)

After filling in all the configuration options via the CLI added in
#3326, we see the options reflected in the UI:


![pr2](https://github.com/oxidecomputer/omicron/assets/1435635/2296874f-f219-445d-9178-d0cf27d143fb)

at which point starting RSS should succeed, and the `Current rack
status: ` line will change to `Initializing`. This will eventually
progress to a new state once RSS completes (whether successful or not).