move techport off bootstrap prefix
The original idea of this PR was to advertise the bootstrap prefix of
the scrimlet through the technician port. However, there are some
concerns with just extending the bootstrap network out the technician
port. This commit is slight modification on the original intent that
still gives us a unique prefix per scrimlet and puts the technician
ports on their own prefix.

We start from the bootstrap prefix and change the leading 16 bits from
fdb0 to fdb1. We keep the subsequent 48 bits the same to maintain a
unique prefix per scrimlet. Then each scrimlet is assigned the address
fdb1:<48-bit-id>::1 which can be used to access wicketd et. al. through
the technician port.