CertificateCreate cert/key Vec<u8> -> String (#3396)
This is intended to address:
https://github.com/oxidecomputer/oxide.rs/issues/161

In short: the API current calls for the key and cert to be specified as
a `Vec<u8>`. While progenitor **could**, say, infer that a `Vec<u8>`
should be some file that we load from disk, in fact the inputs here are
PEM-formatted strings. I think therefore that a `String` is a better way
to model these parameters.