dnswait needs to move to its own package
I originally thought it made sense to put this into "internal-dns" since
it's a straightforward CLI that goes with the library exposed by
"internal-dns".  But it doesn't work due to a few surprising behaviors:

- To ship "dnswait", we'd need to build its parent Rust package as an
  Omicron package.  omicron-package requires that the Omicron package
  name be "internal-dns" if it's going to come from building the
  "internal-dns" Rust package.
- "omicron-package" does some Cargo operation based on packages in
  Cargo.lock.
- Omicron actually has _two_ different "internal-dns" packages in
  Cargo.lock: one from the local workspace, and one from the copy of
  Propolis that pulls in the latest Omicron "main".

As a result, we cannot package up "dnswait" (or any other Rust binary)
if it's in a package in the workspace that's also depended-on by
something else external to Omicron (like Propolis) in the workspace.