ci/tuf-repo: Consistent artifact versions (#3437)
Fixes #3430. Changes:
- Changes the prerelease field in our version from `alpha` to `0.ci`,
which comes before alpha, so that we can use `alpha` or other words
found early in the dictionary and have them count as later than a CI
build.
- Stamps a non-placeholder version into the global zone packages
installed in the host and trampoline OS images.
- Adds `caboose-util`, loosely based on the small part of
hubedit I needed, tweaked for usefulness to a shell script.
- Reads the version out of a Hubris archive's caboose and uses that for
the TUF repository, instead of the control plane version we generate
here.