nexus: provide stable PCI slot assignments for disks (#3408)
Add a `slot` column to the disk table and modify `instance_attach_disk`
to set it when a disk is attached to an instance.

Slots are chosen using the `NextItem` query generator that NICs use for
their slot assignments. The bulk of this change is new code to combine
the `NextItem` fragment with the other SQL statements that mark a disk
as attached (changing its state and setting its `attach_instance_id`) in
a manner that allows the resulting query fragment to be passed as the
body of the SET clause of the UPDATE statement that
`instance_attach_disk` passes to the `attach_resource` generic function.

Slot assignments remain stable for as long as a disk is attached, but
they aren't preserved if a disk is detached and reattached (the
algorithm will choose the first available slot, even if the disk's
previously assigned slot is available).

Tests:

- New integration test
- On a dev cluster, created an instance with a single boot disk, then
attached disks with alphabetically earlier and later names. Verified
that the boot disk remained in slot 0 and that booting the instance
succeeded (i.e. guest firmware was able to find the boot OS
successfully). Also played around with detaching and reattaching disks
and making sure via CRDB shell that their slot assignments were as
expected.

Fixes #3224.