Improve failure mode in omicron-dev tests that fail (#3433)
This fixes a difficult failure mode I ran into recently. Some of the
tests in `omicron-dev` spawn CLI programs, such as `cockroach`. These
are put into a progress group so that, on failure, the entire group can
be signaled and reaped. However, these tests were previously structured
so that the group is only killed on a successful test run; an unwrap or
other panic early would exit before the shutdown code was run. This
leaves the group running and the test hangs.

This PR fixes a couple of those tests, by running the bulk of test code
we care about in a separate fallible future. We run it; store the
result; kill and reap the child processes; and then assert that the body
ran successfully.