Massive change to lrtq FSMs
This changes the structure of the `Fsm` to return a result on API
callbacks. Messages must now be drained after every API call via
`drain_envelopes`. Persistence is indicated implicitly by the type of
ApiOutput returned per API call.

Proptests are also overhauled to have one peer as the system under test
(SUT) and interact solely with that peer via actions. We simulate the
behavior of other peers. There will also be 2 proptests: 1 where the SUT
peer is an initial member of the group, and one where it is a learner.
The initial member test is partially done, but more changes will be
coming in rapidly. I just wanted to get this out for visibility as the
Fsm code itself should be reviewable.