[internal DNS] Implement reqwest::dns::Resolve for internal DNS resolver, use it (#3370)
## Meta

Original title of this PR: "[sled agent] Clarify LazyNexusClient
constructor arguments, cache address"
After feedback, I'm opting to use the internal DNS resolver in the
client which is passed to Progenitor, which changes the scope of this PR
a bit.

# Internal DNS

- Implements
[reqwest::dns::Resolve](https://docs.rs/reqwest/0.11.18/reqwest/dns/trait.Resolve.html)
for the internal DNS server. This enables [usage within the reqwest
client
builder](https://docs.rs/reqwest/0.11.18/reqwest/struct.ClientBuilder.html#method.dns_resolver).
- Add a couple tests verifying that the DNS resolver can successfully be
integrated with Progenitor bindings.

# Sled Agent

- Use the Internal DNS resolver during the construction of the Nexus
client. This *mostly* removes the need for the `LazyNexusClient`
structure which was previously used to lazily perform DNS resolution.
- Replace `LazyNexusClient` with `NexusClientWithResolver`. I would
*prefer* to just have it be a progenitor-created `NexusClient`, but
there are some spots where we're still using the resolver to look up and
pass IP addresses into services which are created by the Sled Agent.
This is worth fixing, but I'm considering it out-of-scope for this PR,
because it'll involve some cross-repo changes.