Idempotently delete volumes during disk delete (#3491)
Previous to this commit, deleting a volume during the disk delete saga
was done by starting the execution of the `volume_delete` saga, but this
is not idempotent. Embed a `volume_delete` subsaga instead.

There's two other sagas where `Nexus::volume_delete` is called: during
the disk creation saga and the snapshot creation saga, both when
unwinding the saga. Unfortunately, embedding a subsaga cannot be done as
part of the *rollback* of a saga, because there's no way to create an
undo node that itself is a subsaga. This commit retains the previous
behaviour during unwind: kick off the `volume_delete` saga. If there's a
failure, then the saga will be parked as "Stuck" anyway.

Fixes #3305