[firewall rules] Fall back to any host or any protocol when filter list is empty (#3511)
The diff is only a couple of lines with [whitespace changes
hidden](https://github.com/oxidecomputer/omicron/pull/3511/files?w=1).

This is one possible approach to addressing #3373. Ports already worked
this way: `None` or empty vec cause the filter to match any port. But
for hosts and and protocols, only `None` falls back to any, while an
empty vec will get passed through and presumably never match anything.

I think even if we end up fixing the API types so that an empty `hosts`
array is rejected with a 400, we still probably want this logic to be
fixed here.