Add `external_dns_ips` RSS parameter (#3573)
Every external DNS IP must be contained in one of the
`internal_services_ip_pool_ranges`, which we validate when we receive
the RSS request in bootstrap agent.

This puts the number of external DNS servers in the control of the user
requesting rack setup; we require at least one but do not put a cap on
it. Should we?

Fixes #3280.