Implement lrtq network layer (#3565)
LRTQ uses persistent TCP connections between peers. These connections
have the
 following properties:

1. Each peer learns of other peer addresses via DDM in the real system.
These
addresses are injected to the peer via
`PeerHandle::load_peer_addresses`.
2. Peers with "higher" IP addresses connect to peers with "lower" ip
addresses - a
 "downhill" strategy
3. Peers maintain persistence via application level `Ping` messages sent
every
second. Peers will timeout after not receiving *any* traffic for 10
seconds.
4. Peers will constantly try to reconnect to any peers they have learned
about with a
 "lower" IP address.

The purpose of these TCP connections is to route LRTQ protocol traffic.
This is
 done in the following manner:

1. When a `TcpStream` is created via a connection from a higher to a
lower peer,
or accepted at a lower peer from a higher peer, the first thing that
happens
 is a handshake.
 2. LRTQ messages may not be sent until the handshake completes.
3. A handshake consists of two messages sent by each side of the
connection:
  a. A `Hello` message used for scheme/version negotiation
b. An `Identify` message used to inform each side of the connection to
its
peer's identity `Baseboard`. For accepting sides to map to the listening
address
of the peer, and not worry about ephemeral ports in testing on a single
machine,
 we also send the canonical IP:Port in the `Identify` message.
 4. Once the handshake completes, LRTQ messages may be sent over the
 "established" connections.
5. All messages except the `Hello` message, which is an exact size, are
framed
 with a 4-byte big-endian size header.
6. LRTQ message sending is triggered by `PeerHandle` method calls which
send
 `PeerApiRequest`s to the underlying "main" peer task.
7. Which messages need to get sent to other peers is determined by the
underlying
 `Fsm`, which is managed by the main peer task.
9. Many API requests, such as `PeerHandle::load_rack_secret`, cannot be
answered
immediately as they require sending messages to other peers and waiting
for
responses. A `responder` is set up for these requests, and when the
`Fsm` returns
a matching output or error, the response message will be sent over the
responder.
 The `PeerHandle` api will block until a response is received.
8. Each connection has its own task and communicates with the main task
over channels.
10. Connection tasks serialize/deserialize and send/receive messages
over the
 established `TcpStream`s.
11. When a message is received on a connection, it is forwarded to the
main peer
 task which calls `Fsm::handle_msg`.
12. When a message needs to be sent out a connection it is sent from the
main
task to that connection managing task, which will serialize and send it
over the network.